# docs/static/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/t/{tenantCode}/admin/users": {
      "get": {
        "operationId": "AdminUserController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "user-1"
                          },
                          "username": {
                            "type": "string",
                            "example": "john"
                          },
                          "email": {
                            "type": "string",
                            "example": "john@example.com",
                            "nullable": true
                          },
                          "emailVerified": {
                            "type": "boolean",
                            "example": false
                          },
                          "phone": {
                            "type": "string",
                            "example": "+821012345678",
                            "nullable": true
                          },
                          "phoneVerified": {
                            "type": "boolean",
                            "example": false
                          },
                          "status": {
                            "type": "string",
                            "example": "ACTIVE"
                          },
                          "mfaEnabled": {
                            "type": "boolean",
                            "example": false
                          },
                          "temporaryPassword": {
                            "type": "boolean",
                            "example": false
                          },
                          "passwordChangeRequired": {
                            "type": "boolean",
                            "example": false
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-05-25T00:00:00.000Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-05-25T00:00:00.000Z"
                          }
                        },
                        "required": [
                          "id",
                          "username",
                          "email",
                          "emailVerified",
                          "phone",
                          "phoneVerified",
                          "status",
                          "mfaEnabled",
                          "temporaryPassword",
                          "passwordChangeRequired",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 20
                    }
                  },
                  "required": ["items", "total", "page", "limit"]
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "List users",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      },
      "post": {
        "operationId": "AdminUserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "1234567890"
                    }
                  },
                  "required": ["id"]
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Create user",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      }
    },
    "/t/{tenantCode}/admin/users/{id}": {
      "get": {
        "operationId": "AdminUserController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "user-1"
                    },
                    "username": {
                      "type": "string",
                      "example": "john"
                    },
                    "email": {
                      "type": "string",
                      "example": "john@example.com",
                      "nullable": true
                    },
                    "emailVerified": {
                      "type": "boolean",
                      "example": false
                    },
                    "phone": {
                      "type": "string",
                      "example": "+821012345678",
                      "nullable": true
                    },
                    "phoneVerified": {
                      "type": "boolean",
                      "example": false
                    },
                    "status": {
                      "type": "string",
                      "example": "ACTIVE"
                    },
                    "mfaEnabled": {
                      "type": "boolean",
                      "example": false
                    },
                    "temporaryPassword": {
                      "type": "boolean",
                      "example": false
                    },
                    "passwordChangeRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-05-25T00:00:00.000Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-05-25T00:00:00.000Z"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "email",
                    "emailVerified",
                    "phone",
                    "phoneVerified",
                    "status",
                    "mfaEnabled",
                    "temporaryPassword",
                    "passwordChangeRequired",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Get user",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      },
      "put": {
        "operationId": "AdminUserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Update user",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      },
      "delete": {
        "operationId": "AdminUserController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Delete user",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      }
    },
    "/t/{tenantCode}/admin/users/{id}/consents": {
      "get": {
        "operationId": "AdminUserController_getConsents",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "consent-1"
                          },
                          "userId": {
                            "type": "string",
                            "example": "user-1"
                          },
                          "clientRefId": {
                            "type": "string",
                            "example": "client-ref-1"
                          },
                          "clientId": {
                            "type": "string",
                            "example": "web-app"
                          },
                          "clientName": {
                            "type": "string",
                            "example": "Web App"
                          },
                          "grantedScopes": {
                            "type": "string",
                            "example": "openid email profile"
                          },
                          "grantedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-05-25T00:00:00.000Z"
                          },
                          "revokedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "example": "ACTIVE"
                          }
                        },
                        "required": [
                          "id",
                          "userId",
                          "clientRefId",
                          "clientId",
                          "clientName",
                          "grantedScopes",
                          "grantedAt",
                          "revokedAt",
                          "status"
                        ]
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 20
                    }
                  },
                  "required": ["items", "total", "page", "limit"]
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "List user consents",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      }
    },
    "/t/{tenantCode}/admin/users/{id}/consents/history": {
      "get": {
        "operationId": "AdminUserController_getConsentHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "consent-1"
                          },
                          "userId": {
                            "type": "string",
                            "example": "user-1"
                          },
                          "clientRefId": {
                            "type": "string",
                            "example": "client-ref-1"
                          },
                          "clientId": {
                            "type": "string",
                            "example": "web-app"
                          },
                          "clientName": {
                            "type": "string",
                            "example": "Web App"
                          },
                          "grantedScopes": {
                            "type": "string",
                            "example": "openid email profile"
                          },
                          "grantedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-05-25T00:00:00.000Z"
                          },
                          "revokedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "example": "ACTIVE"
                          }
                        },
                        "required": [
                          "id",
                          "userId",
                          "clientRefId",
                          "clientId",
                          "clientName",
                          "grantedScopes",
                          "grantedAt",
                          "revokedAt",
                          "status"
                        ]
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 20
                    }
                  },
                  "required": ["items", "total", "page", "limit"]
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "List user consent history",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      }
    },
    "/t/{tenantCode}/admin/users/{id}/roles": {
      "get": {
        "operationId": "AdminUserController_getRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "role-1"
                      },
                      "code": {
                        "type": "string",
                        "example": "USER_ADMIN"
                      },
                      "name": {
                        "type": "string",
                        "example": "User Admin"
                      },
                      "description": {
                        "type": "string",
                        "example": "Manage users",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-25T00:00:00.000Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-25T00:00:00.000Z"
                      }
                    },
                    "required": [
                      "id",
                      "code",
                      "name",
                      "description",
                      "createdAt",
                      "updatedAt"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "List user roles",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      }
    },
    "/t/{tenantCode}/admin/users/{id}/roles/{roleId}": {
      "post": {
        "operationId": "AdminUserController_assignRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Assign role to user",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      },
      "delete": {
        "operationId": "AdminUserController_removeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Remove role from user",
        "tags": ["Admin Users"],
        "description": "User management. Passwords are hashed, sensitive credentials are not returned, status changes are audited, and personally identifiable data is kept minimal."
      }
    },
...
  },
  "info": {
    "title": "Auth API",
    "description": "OIDC Authorization Server API Reference",
    "version": "0.1.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Service origin"
    },
    {
      "url": "/api",
      "description": "Vite proxy"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}